+++ tests/train/test_bc_warm_start_entry.py
@@ -206,8 +206,7 @@ def test_a_cpu_smoke_plays_one_legal_game_from_the_warm_started_net(tmp_path: Pa
         net, torch.device("cpu"), encoding_spec=spec,
         fused_graph_caps=FusedGraphCapsSpec(max_fused_edges=57149441, max_fused_nodes=1785921),
         inference_batching=InferenceBatchingSpec(inference_batch_size=8, inference_max_wait_ms=10),
-        max_in_flight=4, amp_dtype="bf16",
-    )
+        max_in_flight=4, )
     try:
         def _player() -> Any:
             # `n_sims` deliberately small: this is a liveness smoke on the default tier, not a
